+++ lib/wallet.js
@@ -4,6 +4,7 @@ const scrypt = require('scrypt-async');
 const uuid = require('uuid');
 const bip39 = require('bip39');
 const wordlist = bip39.wordlists.EN;
+const util = require('./util');
 const wallet = {};
 
 let keystore = function (privateKey) {
@@ -15,16 +16,6 @@ let keystore = function (privateKey) {
   }
   this._privateKey = privateKey;
   this._publicKey = ethUtil.privateToPublic(privateKey);
-};
-
-bufferToArray = function (buffer) {
-  const length = buffer.length;
-  let data = [];
-
-  for (let i=0; i<length; i++) {
-    data.push(buffer[i]);
-  }
-  return data;
 };
 
 wallet.validateSeed = function (seed) {
@@ -93,7 +84,7 @@ wallet.fromV3 = function (v3, password, callback) {
     if (v3.crypto.kdf === 'scrypt') {
       const kdfparams = v3.crypto.kdfparams;
       const saltBuffer = new Buffer(kdfparams.salt, 'hex');
-      const salt = bufferToArray(saltBuffer);
+      const salt = util.bufferToArray(saltBuffer);
 
       scrypt(password, salt, {N: kdfparams.n, r: kdfparams.r, p: kdfparams.p, dklen: kdfparams.dklen, encoding: 'binary'}, cb);
     } else if (v3.crypto.kdf === 'pbkdf2') {
@@ -243,7 +234,7 @@ keystore.prototype.toV3 = function (password, options, callback) {
         cb(derivedKey);
       });
     } else if (kdf === 'scrypt') {
-      const saltUse = bufferToArray(salt);
+      const saltUse = util.bufferToArray(salt);
 
       kdfparams.n = options.n || 262144;
       kdfparams.r = options.r || 8;
